refactor Flag constants

+++ src/ContractGenerator/ContractEventClassGenerator.cs
@@ -10,7 +10,7 @@ public class ContractEventClassGenerator
     /// Generate will produce a chunk of C# code that serves as the Event class of the AElf Contract. Events are used internally to represent events that have happened during the execution of a smart contract.
     /// </summary>
     //TODO Implement following https://github.com/AElfProject/contract-plugin/blob/453bebfec0dd2fdcc06d86037055c80721d24e8a/src/contract_csharp_generator.cc#L546
-    public string Generate(MessageDescriptor descriptor, uint flags)
+    public string Generate(MessageDescriptor descriptor, byte flags)
     {
         // descriptor.Fields[0].ContainingType.Name;
         throw new NotImplementedException();
